desktop/requirements-dev.txt
-r requirements.txt
alembic==1.10.4
celery-types==0.22.0
coverage==7.2.2
flake8==6.1.0
iniconfig==1.1.1
Mako==1.2.3
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.5.1
mypy-extensions==1.0.0
pluggy==1.4.0
psutil==5.9.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyQt5-stubs==5.15.6.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-env==1.1.3
pytest-mock==3.10.0
pytest-qt==4.4.0
pytest-sqlalchemy-mock==0.1.5
python-dateutil==2.8.2
PyYAML==6.0.1
toml==0.10.2
watchdog==2.1.6
